+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/robots/goldenduck/Claw.java
@@ -1,0 +1,4 @@
+package org.firstinspires.ftc.teamcode.robots.goldenduck;
+
+public class Claw {
+}
